storage: add filesystem uuid to BlockDev using blkid
The BlockDev struct is extended by the filesystem uuid of the respective
partition, if any. This is done by executing blkid.
The filesystme uuid is used by grub to identify partitions.

IMPORTANT: The blkid binary needs to be added to the initramfs by u-root
using the -file flag.

TODO: The filesystem uuid schould be extracted from the divece directly in
go instead of calling bkid.