grub: add efi support in grub.cfg parser
When parsing the grub.cfg, now it is regognized when root is set to an other
partition in an bootconfig entry. So in efi systems where the grub.cfg
resides on a dedicated efi partiton whereas the kernel my be on other
partition the path is set correctly to load the kernel.